Criação de uma verificação para atualizar ou deletar um post, sendo que apenas o user proprietario tem essa permissão

+++ Controllers/V1/PostsController.cs
@@ -7,6 +7,7 @@ using Tweetbook.Contracts.V1;
 using Tweetbook.Contracts.V1.Request;
 using Tweetbook.Contracts.V1.Response;
 using Tweetbook.Domain;
+using Tweetbook.Extensions;
 using Tweetbook.Services;
 
 namespace Tweetbook.Controllers.V1
@@ -29,7 +30,7 @@ namespace Tweetbook.Controllers.V1
         }
 
         [HttpGet(ApiRoutes.Posts.Get)]
-        public async Task<IActionResult> GetAll([FromRoute] Guid postId)
+        public async Task<IActionResult> Get([FromRoute] Guid postId)
         {
             var post = await _postService.GetPostByIdAsync(postId);
 
@@ -42,7 +43,10 @@ namespace Tweetbook.Controllers.V1
         [HttpPost(ApiRoutes.Posts.Create)]
         public async Task<IActionResult> Create([FromBody] CreatePostRequest postRequest)
         {
-            var post = new Post { Name = postRequest.Name };
+            var post = new Post
+            {
+                Name = postRequest.Name
+            };
 
             await _postService.CreatePostAsync(post);
 
@@ -55,13 +59,16 @@ namespace Tweetbook.Controllers.V1
         }
 
         [HttpPut(ApiRoutes.Posts.Update)]
-        public async Task<IActionResult> GetAll([FromRoute] Guid postId, [FromBody] UpdatePostRequest postRequest)
+        public async Task<IActionResult> Update([FromRoute] Guid postId, [FromBody] UpdatePostRequest postRequest)
         {
-            var post = new Post
-            {
-                Id = postId,
-                Name = postRequest.Name
-            };
+            var userOwnsPost = await _postService.UserOwnsPostAsync(postId, HttpContext.GetUserId());
+
+            if (!userOwnsPost)
+                return BadRequest(new { 
+                    error = "You do not own this post" });
+
+            var post = await _postService.GetPostByIdAsync(postId);
+            post.Name = postRequest.Name;
 
             var updated = await _postService.UpdatePostAsync(post);
 
@@ -74,6 +81,14 @@ namespace Tweetbook.Controllers.V1
         [HttpDelete(ApiRoutes.Posts.Delete)]
         public async Task<IActionResult> Delete([FromRoute] Guid postId)
         {
+            var userOwnsPost = await _postService.UserOwnsPostAsync(postId, HttpContext.GetUserId());
+
+            if (!userOwnsPost)
+                return BadRequest(new
+                {
+                    error = "You do not own this post"
+                });
+
             var deleted = await _postService.DeletePostAsync(postId);
 
             if (deleted)
